Start restructuring
Known issues: A lot of things still need to be adjusted.
LP files for the tests are among them.
