+++ src/oemof/solph/__init__.py
@@ -1,23 +1,22 @@
 __version__ = "0.4.5.dev0"
 
 from . import constraints  # noqa: F401
-from . import custom  # noqa: F401
+from . import experimental  # noqa: F401
 from . import helpers  # noqa: F401
+from . import processing  # noqa: F401
 from . import views  # noqa: F401
-from .components import ExtractionTurbineCHP  # noqa: F401
-from .components import GenericCHP  # noqa: F401
-from .components import GenericStorage  # noqa: F401
-from .components import OffsetTransformer  # noqa: F401
-from .groupings import GROUPINGS  # noqa: F401
-from .models import Model  # noqa: F401
-from .network import Bus  # noqa: F401
-from .network import EnergySystem  # noqa: F401
-from .network import Flow  # noqa: F401
-from .network import Sink  # noqa: F401
-from .network import Source  # noqa: F401
-from .network import Transformer  # noqa: F401
-from .options import Investment  # noqa: F401
-from .options import NonConvex  # noqa: F401
-from .plumbing import sequence  # noqa: F401
-from .processing import parameter_as_dict  # noqa: F401
-from .processing import results  # noqa: F401
+from ._buses import Bus  # noqa: F401
+from ._components import ExtractionTurbineCHP  # noqa: F401
+from ._components import GenericCHP  # noqa: F401
+from ._components import GenericStorage  # noqa: F401
+from ._components import OffsetTransformer  # noqa: F401
+from ._components import Sink  # noqa: F401
+from ._components import Source  # noqa: F401
+from ._components import Transformer  # noqa: F401
+from ._energy_system import EnergySystem  # noqa: F401
+from ._flows import Flow  # noqa: F401
+from ._groupings import GROUPINGS  # noqa: F401
+from ._models import Model  # noqa: F401
+from ._options import Investment  # noqa: F401
+from ._options import NonConvex  # noqa: F401
+from ._plumbing import sequence  # noqa: F401
