+++ src/oemof/solph/_buses/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+"""
+This module is designed to hold busses with their classes and
+associated individual constraints (blocks) and groupings. Therefore this
+module holds the class definition and the block directly located by each other.
+
+Note that only mature code is imported,
+experimental code should be included in oemof.experimental.
+"""
+
+from .bus import Bus  # noqa: F401
