+++ src/oemof/solph/_components/__init__.py
@@ -4,9 +4,15 @@
 This module is designed to hold components with their classes and
 associated individual constraints (blocks) and groupings. Therefore this
 module holds the class definition and the block directly located by each other.
+
+Note that only mature code is imported,
+experimental code should be included in oemof.experimental.
 """
 
 from .extraction_turbine_chp import ExtractionTurbineCHP  # noqa: F401
 from .generic_chp import GenericCHP  # noqa: F401
 from .generic_storage import GenericStorage  # noqa: F401
 from .offset_transformer import OffsetTransformer  # noqa: F401
+from .sink import Sink
+from .source import Source
+from .transformer import Transformer
