+++ src/oemof/solph/_components/sink.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-solph version of oemof.network.Source
+solph version of oemof.network.Sink
 
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
 SPDX-FileCopyrightText: Simon Hilpert
@@ -15,15 +15,15 @@ SPDX-License-Identifier: MIT
 
 from oemof.network import network as on
 
-from ._helpers import check_node_object_for_missing_attribute
+from oemof.solph._helpers import check_node_object_for_missing_attribute
 
 
-class Source(on.Source):
-    """An object with one output flow."""
+class Sink(on.Sink):
+    """An object with one input flow."""
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        check_node_object_for_missing_attribute(self, "outputs")
+        check_node_object_for_missing_attribute(self, "inputs")
 
     def constraint_group(self):
         pass
