+++ src/oemof/solph/_components/source.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-solph version of oemof.network.Sink
+solph version of oemof.network.Source
 
 SPDX-FileCopyrightText: Uwe Krien <krien@uni-bremen.de>
 SPDX-FileCopyrightText: Simon Hilpert
@@ -15,15 +15,15 @@ SPDX-License-Identifier: MIT
 
 from oemof.network import network as on
 
-from ._helpers import check_node_object_for_missing_attribute
+from oemof.solph._helpers import check_node_object_for_missing_attribute
 
 
-class Sink(on.Sink):
-    """An object with one input flow."""
+class Source(on.Source):
+    """An object with one output flow."""
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        check_node_object_for_missing_attribute(self, "inputs")
+        check_node_object_for_missing_attribute(self, "outputs")
 
     def constraint_group(self):
         pass
