--- src/oemof/solph/blocks/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Creating sets, variables, constraints and parts of the objective function
-for the specified groups.
-"""
-
-from .bus import Bus  # noqa: F401
-from .flow import Flow  # noqa: F401
-from .investment_flow import InvestmentFlow  # noqa: F401
-from .non_convex_flow import NonConvexFlow  # noqa: F401
-from .transformer import Transformer  # noqa: F401
