--- src/oemof/solph/custom/__init__.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-This module is designed to hold in-development components with their classes
-and associated individual constraints (blocks) and groupings.
-
-Requirements for documentation and unit tests are relaxed,
-so code within this module might not have production quality.
-"""
-
-from .electrical_line import ElectricalBus  # noqa: F401
-from .electrical_line import ElectricalLine  # noqa: F401
-from .generic_caes import GenericCAES  # noqa: F401
-from .link import Link  # noqa: F401
-from .piecewise_linear_transformer import (  # noqa: F401
-    PiecewiseLinearTransformer,
-)
-from .sink_dsm import SinkDSM  # noqa: F401
