+++ src/oemof/solph/experimental.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+"""
+This module is designed to hold in-development components with their classes
+and associated individual constraints (blocks) and groupings.
+
+Requirements for documentation and unit tests are relaxed,
+so code referred to within this module might not have production quality.
+"""
+
+from ._buses.electrical_bus import ElectricalBus  # noqa: F401
+from ._components.generic_caes import GenericCAES  # noqa: F401
+from ._components.link import Link  # noqa: F401
+from ._components.piecewise_linear_transformer import (  # noqa: F401
+    PiecewiseLinearTransformer,
+)
+from ._components.sink_dsm import SinkDSM  # noqa: F401
+from ._flows.electrical_line import ElectricalLine  # noqa: F401
