--- src/oemof/solph/network/__init__.py
@@ -1,18 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""
-Classes used to model energy supply systems within solph.
-
-Classes are derived from oemof core network classes and adapted for specific
-optimization tasks. An energy system is modelled as a graph/network of nodes
-with very specific constraints on which types of nodes are allowed to be
-connected.
-"""
-
-from ._helpers import check_node_object_for_missing_attribute  # noqa: F401
-from .bus import Bus  # noqa: F401
-from .energy_system import EnergySystem  # noqa: F401
-from .flow import Flow  # noqa: F401
-from .sink import Sink  # noqa: F401
-from .source import Source  # noqa: F401
-from .transformer import Transformer  # noqa: F401
