+++ tests/lp_files/linear_transformer.lp
@@ -8,41 +8,41 @@ objective:
 
 s.t.
 
-c_e_Bus_balance(electricity_0)_:
+c_e_BusBlock_balance(electricity_0)_:
 +1 flow(powerplantGas_electricity_0)
 = 0
 
-c_e_Bus_balance(electricity_1)_:
+c_e_BusBlock_balance(electricity_1)_:
 +1 flow(powerplantGas_electricity_1)
 = 0
 
-c_e_Bus_balance(electricity_2)_:
+c_e_BusBlock_balance(electricity_2)_:
 +1 flow(powerplantGas_electricity_2)
 = 0
 
-c_e_Bus_balance(gas_0)_:
+c_e_BusBlock_balance(gas_0)_:
 +1 flow(gas_powerplantGas_0)
 = 0
 
-c_e_Bus_balance(gas_1)_:
+c_e_BusBlock_balance(gas_1)_:
 +1 flow(gas_powerplantGas_1)
 = 0
 
-c_e_Bus_balance(gas_2)_:
+c_e_BusBlock_balance(gas_2)_:
 +1 flow(gas_powerplantGas_2)
 = 0
 
-c_e_Transformer_relation(powerplantGas_gas_electricity_0)_:
+c_e_TransformerBlock_relation(powerplantGas_gas_electricity_0)_:
 +0.57999999999999996 flow(gas_powerplantGas_0)
 -1 flow(powerplantGas_electricity_0)
 = 0
 
-c_e_Transformer_relation(powerplantGas_gas_electricity_1)_:
+c_e_TransformerBlock_relation(powerplantGas_gas_electricity_1)_:
 +0.57999999999999996 flow(gas_powerplantGas_1)
 -1 flow(powerplantGas_electricity_1)
 = 0
 
-c_e_Transformer_relation(powerplantGas_gas_electricity_2)_:
+c_e_TransformerBlock_relation(powerplantGas_gas_electricity_2)_:
 +0.57999999999999996 flow(gas_powerplantGas_2)
 -1 flow(powerplantGas_electricity_2)
 = 0
