+++ tests/test_components.py
@@ -18,7 +18,7 @@ from oemof.solph import Bus
 from oemof.solph import Flow
 from oemof.solph import Investment
 from oemof.solph import NonConvex
-from oemof.solph import components
+from oemof.solph import _components as components
 
 # ********* GenericStorage *********
