+++ tests/test_constraints_module.py
@@ -50,7 +50,7 @@ def test_something_else():
     line21 = energysystem.groups["powerline_2_1"]
     solph.constraints.equate_variables(
         om,
-        om.InvestmentFlow.invest[line12, bel2],
-        om.InvestmentFlow.invest[line21, bel1],
+        om.InvestmentFlowBlock.invest[line12, bel2],
+        om.InvestmentFlowBlock.invest[line21, bel1],
         name="my_name",
     )
