+++ tests/test_models.py
@@ -21,7 +21,7 @@ from oemof.solph.helpers import calculate_timeincrement
 def test_timeincrement_with_valid_timeindex():
     datetimeindex = pd.date_range("1/1/2012", periods=1, freq="H")
     es = solph.EnergySystem(timeindex=datetimeindex)
-    m = solph.models.BaseModel(es)
+    m = solph._models.BaseModel(es)
     assert m.timeincrement[0] == 1
     assert es.timeindex.freq.nanos / 3.6e12 == 1
 
@@ -29,18 +29,18 @@ def test_timeincrement_with_valid_timeindex():
 def test_timeincrement_with_non_valid_timeindex():
     with pytest.raises(AttributeError):
         es = solph.EnergySystem(timeindex=4)
-        solph.models.BaseModel(es)
+        solph._models.BaseModel(es)
 
 
 def test_timeincrement_value():
     es = solph.EnergySystem(timeindex=4)
-    m = solph.models.BaseModel(es, timeincrement=3)
+    m = solph._models.BaseModel(es, timeincrement=3)
     assert m.timeincrement[0] == 3
 
 
 def test_timeincrement_list():
     es = solph.EnergySystem(timeindex=4)
-    m = solph.models.BaseModel(es, timeincrement=[0, 1, 2, 3])
+    m = solph._models.BaseModel(es, timeincrement=[0, 1, 2, 3])
     assert m.timeincrement[3] == 3
 
 
@@ -96,7 +96,7 @@ def test_optimal_solution():
     es.add(bel)
     es.add(solph.Sink(inputs={bel: solph.Flow(nominal_value=5, fix=[1])}))
     es.add(solph.Source(outputs={bel: solph.Flow(variable_costs=5)}))
-    m = solph.models.Model(es, timeincrement=1)
+    m = solph.Model(es, timeincrement=1)
     m.solve("cbc")
     m.results()
     solph.processing.meta_results(m)
@@ -118,7 +118,7 @@ def test_infeasible_model():
                     }
                 )
             )
-            m = solph.models.Model(es, timeincrement=1)
+            m = solph.Model(es, timeincrement=1)
             m.solve(solver="cbc")
             assert "Optimization ended with status" in str(w[0].message)
             solph.processing.meta_results(m)
