+++ tests/test_processing.py
@@ -19,13 +19,13 @@ from pandas.testing import assert_series_equal
 from oemof.solph import Bus
 from oemof.solph import EnergySystem
 from oemof.solph import Flow
+from oemof.solph import GenericStorage
 from oemof.solph import Investment
 from oemof.solph import Model
 from oemof.solph import Sink
 from oemof.solph import Transformer
 from oemof.solph import processing
 from oemof.solph import views
-from oemof.solph.components import GenericStorage
 
 
 class TestParameterResult:
