+++ tests/test_scripts/test_solph/test_lopf/test_lopf.py
@@ -23,7 +23,7 @@ from oemof.solph import Investment
 from oemof.solph import Model
 from oemof.solph import Sink
 from oemof.solph import Source
-from oemof.solph import custom
+from oemof.solph import experimental
 from oemof.solph import processing
 from oemof.solph import views
 
@@ -41,16 +41,16 @@ def test_lopf(solver="cbc"):
 
     logging.info("Create oemof.solph objects")
 
-    b_el0 = custom.ElectricalBus(label="b_0", v_min=-1, v_max=1)
+    b_el0 = experimental.ElectricalBus(label="b_0", v_min=-1, v_max=1)
 
-    b_el1 = custom.ElectricalBus(label="b_1", v_min=-1, v_max=1)
+    b_el1 = experimental.ElectricalBus(label="b_1", v_min=-1, v_max=1)
 
-    b_el2 = custom.ElectricalBus(label="b_2", v_min=-1, v_max=1)
+    b_el2 = experimental.ElectricalBus(label="b_2", v_min=-1, v_max=1)
 
     es.add(b_el0, b_el1, b_el2)
 
     es.add(
-        custom.ElectricalLine(
+        experimental.ElectricalLine(
             input=b_el0,
             output=b_el1,
             reactance=0.0001,
@@ -61,7 +61,7 @@ def test_lopf(solver="cbc"):
     )
 
     es.add(
-        custom.ElectricalLine(
+        experimental.ElectricalLine(
             input=b_el1,
             output=b_el2,
             reactance=0.0001,
@@ -72,7 +72,7 @@ def test_lopf(solver="cbc"):
     )
 
     es.add(
-        custom.ElectricalLine(
+        experimental.ElectricalLine(
             input=b_el2,
             output=b_el0,
             reactance=0.0001,
