+++ tests/test_scripts/test_solph/test_storage_investment/test_invest_storage_regression.py
@@ -54,7 +54,7 @@ def test_regression_investment_storage(solver="cbc"):
     )
 
     # Investment storage
-    solph.components.GenericStorage(
+    solph.GenericStorage(
         label="storage",
         inputs={
             bel: solph.Flow(
